InputRow component
