+++ src/common/components/inputRow/inputRow.module.scss
@@ -1,0 +1,10 @@
+.inputRow {
+  display: grid;
+  grid-template-columns: 2fr 1fr;
+
+  .inputColumn {
+  }
+
+  .infoColumn {
+  }
+}
