+++ src/domain/event/formSections/typeSection/typeSection.module.scss
@@ -10,6 +10,6 @@
   }
 
   @include medium_up {
-    grid-template-columns: 1fr 1fr;
+    grid-template-columns: 1fr 1fr 1fr;
   }
 }
